LayoutTests/platform/ios-sim-deprecated/fast/dom/HTMLTemplateElement/cycles-in-shadow-expected.txt

CONSOLE MESSAGE: line 12: TypeError: undefined is not a function (evaluating 'div.webkitCreateShadowRoot()')
Test that cycle detection traverses over both templates and shadow roots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
